samples: net: http_server: use ALPN for browser HTTP2 compatibilty
Update HTTP server sample to enable ALPN, so that web browsers can use
HTTP2 protocol when communicating with the server.

When using HTTP2, browsers are allowed to treat certain cipher suites as
a connection error of type INADEQUATE_SECURITY - see RFC 7540 Appendix
A. In practice this seems to be the case for Chrome and Firefox. So the
sample's cipher suite configuration and certificates are updated to
use TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256, and add the default net
sample IP address as a subjectAltName.

The new CA certificate in PEM format is retained, so that this can be
imported into the browser certificate list for testing purposes if
desired.

Signed-off-by: Matt Rodgers <mrodgers@witekio.com>